Opdracht afgerond Pizzacalculator
..

+++ module1/deel1/speelhal.py
@@ -1,4 +1,4 @@
-vip = int(7.45 * 4)
-perpersoon = int((9 * 0.37) * 4)
+vip = float(7.45 * 4)
+perpersoon = float((9 * 0.37) * 4)
 
 print(vip + perpersoon)
